@@ -1,11 +1,12 @@
-﻿
-namespace Frost.Net.Models
+﻿namespace Frost.Net.Models;
+
+/// <summary>
+/// CMYK Color model - Cyan, Magenta, Yellow, Key
+/// </summary>
+public struct CMYK
 {
-    public struct CMYK
-    {
-        public float c;
-        public float m;
-        public float y;
-        public float k;
-    }
+    public float c;
+    public float m;
+    public float y;
+    public float k;
 }
