@@ -1,138 +1,141 @@
 ﻿using System.Runtime.CompilerServices;
+using Frost.Net.Interoperability;
 
-namespace Frost.Net.Models
+namespace Frost.Net.Models;
+
+/// <summary>
+/// High dynamic range RGBA Color model - Red, Green, Blue, Alpha
+/// </summary>
+public struct HDRColor
 {
-    public struct HDRColor
-    {
-        public float r;
-        public float g;
-        public float b;
-        public float a;
+    public float r;
+    public float g;
+    public float b;
+    public float a;
 
-        public HDRColor(float r, float g, float b, float a)
+    public HDRColor(float r, float g, float b, float a)
+    {
+        this.r = r;
+        this.g = g;
+        this.b = b;
+        this.a = a;
+    }
+    public HDRColor(int value)
+    {
+        Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
+        Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
+        Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
+        Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
+        unsafe
         {
-            this.r = r;
-            this.g = g;
-            this.b = b;
-            this.a = a;
+            fixed (void* output = &r)
+                FrostApi.Color.Rgba8ToRgba32(value, new IntPtr(output));
         }
-        public HDRColor(int value)
+    }
+    public HDRColor(Color rgba)
+    {
+        unsafe
         {
             Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
             Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
             Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
             Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
-            unsafe
-            {
-                fixed (void* output = &r)
-                    FrostApi.Color.Rgba8ToRgba32(value, new IntPtr(output));
-            }
-        }
-        public HDRColor(Color rgba)
-        {
-            unsafe
-            {
-                Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
 
-                fixed (void* output = &r)
-                    FrostApi.Color.Rgba8ToRgba32(rgba.value, new IntPtr(output));
-            }
+            fixed (void* output = &r)
+                FrostApi.Color.Rgba8ToRgba32(rgba.value, new IntPtr(output));
         }
-        public HDRColor(HSVA hsva)
+    }
+    public HDRColor(HSVA hsva)
+    {
+        unsafe
         {
-            unsafe
-            {
-                Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
 
-                fixed (void* output = &r)
-                    FrostApi.Color.HsvaToRgba32(new IntPtr(&hsva.h), new IntPtr(output));
-            }
+            fixed (void* output = &r)
+                FrostApi.Color.HsvaToRgba32(new IntPtr(&hsva.h), new IntPtr(output));
         }
-        public HDRColor(HSLA hsla)
+    }
+    public HDRColor(HSLA hsla)
+    {
+        unsafe
         {
-            unsafe
-            {
-                Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
 
-                fixed (void* output = &r)
-                    FrostApi.Color.HslaToRgba32(new IntPtr(&hsla.h), new IntPtr(output));
-            }
+            fixed (void* output = &r)
+                FrostApi.Color.HslaToRgba32(new IntPtr(&hsla.h), new IntPtr(output));
         }
-        public HDRColor(CMYK cmyk)
+    }
+    public HDRColor(CMYK cmyk)
+    {
+        unsafe
         {
-            unsafe
-            {
-                Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
-                Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out r); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out g); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out b); // Skip initialization, initialized with Interop
+            Unsafe.SkipInit(out a); // Skip initialization, initialized with Interop
 
-                fixed (void* output = &r)
-                    FrostApi.Color.CmykToRgba32(new IntPtr(&cmyk.c), new IntPtr(output));
-            }
+            fixed (void* output = &r)
+                FrostApi.Color.CmykToRgba32(new IntPtr(&cmyk.c), new IntPtr(output));
         }
+    }
 
-        public HSVA ToHsva()
+    public HSVA ToHsva()
+    {
+        unsafe
         {
-            unsafe
-            {
-                float* hsva = stackalloc float[4];
-                fixed (void* input = &r)
-                    FrostApi.Color.Rgba32ToHsva(new IntPtr(input), new IntPtr(hsva));
-                return new HSVA() { h = hsva[0], s = hsva[1], v = hsva[2], a = hsva[3] };
-            }
+            float* hsva = stackalloc float[4];
+            fixed (void* input = &r)
+                FrostApi.Color.Rgba32ToHsva(new IntPtr(input), new IntPtr(hsva));
+            return new HSVA() { h = hsva[0], s = hsva[1], v = hsva[2], a = hsva[3] };
         }
-        public HSLA ToHsla()
+    }
+    public HSLA ToHsla()
+    {
+        unsafe
         {
-            unsafe
-            {
-                float* hsla = stackalloc float[4];
-                fixed (void* input = &r)
-                    FrostApi.Color.Rgba32ToHsla(new IntPtr(input), new IntPtr(hsla));
-                return new HSLA() { h = hsla[0], s = hsla[1], l = hsla[2], a = hsla[3] };
-            }
+            float* hsla = stackalloc float[4];
+            fixed (void* input = &r)
+                FrostApi.Color.Rgba32ToHsla(new IntPtr(input), new IntPtr(hsla));
+            return new HSLA() { h = hsla[0], s = hsla[1], l = hsla[2], a = hsla[3] };
         }
-        public CMYK ToCmyk()
+    }
+    public CMYK ToCmyk()
+    {
+        unsafe
         {
-            unsafe
-            {
-                float* cmyk = stackalloc float[4];
-                fixed (void* input = &r)
-                    FrostApi.Color.Rgba32ToCmyk(new IntPtr(input), new IntPtr(cmyk));
-                return new CMYK() { c = cmyk[0], m = cmyk[1], y = cmyk[2], k = cmyk[3] };
-            }
+            float* cmyk = stackalloc float[4];
+            fixed (void* input = &r)
+                FrostApi.Color.Rgba32ToCmyk(new IntPtr(input), new IntPtr(cmyk));
+            return new CMYK() { c = cmyk[0], m = cmyk[1], y = cmyk[2], k = cmyk[3] };
         }
+    }
 
 
-        public static HDRColor FromHsva(HSVA hsva)
-        {
-            HDRColor result;
-            Unsafe.SkipInit(out result);
-            unsafe { FrostApi.Color.HsvaToRgba32(new IntPtr(&hsva.h), new IntPtr(&result.r)); }
-            return result;
-        }
-        public static HDRColor FromHsla(HSLA hsla)
-        {
-            HDRColor result;
-            Unsafe.SkipInit(out result);
-            unsafe { FrostApi.Color.HslaToRgba32(new IntPtr(&hsla.h), new IntPtr(&result.r)); }
-            return result;
-        }
-        public static HDRColor FromCmyk(CMYK cmyk)
-        {
-            HDRColor result;
-            Unsafe.SkipInit(out result);
-            unsafe { FrostApi.Color.CmykToRgba32(new IntPtr(&cmyk.c), new IntPtr(&result.r)); }
-            return result;
-        }
+    public static HDRColor FromHsva(HSVA hsva)
+    {
+        HDRColor result;
+        Unsafe.SkipInit(out result);
+        unsafe { FrostApi.Color.HsvaToRgba32(new IntPtr(&hsva.h), new IntPtr(&result.r)); }
+        return result;
+    }
+    public static HDRColor FromHsla(HSLA hsla)
+    {
+        HDRColor result;
+        Unsafe.SkipInit(out result);
+        unsafe { FrostApi.Color.HslaToRgba32(new IntPtr(&hsla.h), new IntPtr(&result.r)); }
+        return result;
+    }
+    public static HDRColor FromCmyk(CMYK cmyk)
+    {
+        HDRColor result;
+        Unsafe.SkipInit(out result);
+        unsafe { FrostApi.Color.CmykToRgba32(new IntPtr(&cmyk.c), new IntPtr(&result.r)); }
+        return result;
     }
 }
