@@ -1,10 +1,12 @@
-﻿namespace Frost.Net.Models
+﻿namespace Frost.Net.Models;
+
+/// <summary>
+/// HSLA Color model - Hue, Saturation, Lightness, Alpha
+/// </summary>
+public struct HSLA
 {
-    public struct HSLA
-    {
-        public float h;
-        public float s;
-        public float l;
-        public float a;
-    }
+    public float h;
+    public float s;
+    public float l;
+    public float a;
 }
