@@ -1,10 +1,12 @@
-﻿namespace Frost.Net.Models
+﻿namespace Frost.Net.Models;
+
+/// <summary>
+/// HSVA Color model - Hue, Saturation, Value, Alpha
+/// </summary>
+public struct HSVA
 {
-    public struct HSVA
-    {
-        public float h;
-        public float s;
-        public float v;
-        public float a;
-    }
+    public float h;
+    public float s;
+    public float v;
+    public float a;
 }
