@@ -1,7 +1,16 @@
-﻿Frost.Net.Random r = new Frost.Net.Random();
+﻿using Frost.Net;
+var mainThread = Frost.Net.Thread.ThisThread;
+var e = new Frost.Net.Synchronization.SyncEvent();
+var th = new Frost.Net.Thread(() => {
+	e.Wait();
+	while(true)
+		Frost.Net.Thread.Message.Send(mainThread, () => { Console.WriteLine("Hello, World!"); });
+});
+var message = new Frost.Net.Thread.Message();
+e.Signal();
 while (true)
 {
-	Console.WriteLine(r.NextF64(5, 5));
-	Thread.Sleep(10);
+	message.Wait();
+	message.Dispatch();
 }
 
