@@ -1,5 +1,5 @@
 #include "../semantic_version.api.hpp"
-bool semantic_version_check_compatibility(semantic_version version, semantic_version target)
+FROST_API bool _stdcall frost_api_semantic_version_check_compatibility(semantic_version version, semantic_version target)
 {
 	return
 		(version.major_version == target.major_version) &&
@@ -8,7 +8,7 @@ bool semantic_version_check_compatibility(semantic_version version, semantic_ver
 			((version.minor_version == target.minor_version) && (version.patch >= target.patch))
 		);
 }
-void semantic_version_get_api_version(semantic_version* output)
+FROST_API void _stdcall frost_api_semantic_version_get_api_version(semantic_version* output)
 {
 	output->major_version	= 0;
 	output->minor_version	= 12;
