@@ -1,14 +1,32 @@
+#include <typeinfo>
+#include <string>
+#include <exception>
+#include <stdio.h>
 #include "../resource.impl.hpp"
-u64 frost::impl::resource::impl_get_reference_count() const
+#include "../debug.impl.hpp"
+namespace frost::api
 {
-	return this->_reference_count;
-}
-void frost::impl::resource::impl_acquire_reference() const
-{
-	this->_reference_count.fetch_add(1); 
-}
-void frost::impl::resource::impl_release_reference() const
-{
-	if (_reference_count.fetch_sub(1) == 1)
-		delete this;
+	resource::~resource()
+	{
+		frost::impl::debug::log_resource_destruction(this, typeid(*this));
+	}
+
+	resource::resource() : _reference_count()
+	{
+		frost::impl::debug::log_resource_creation(this);
+	}
+
+	u64 resource::get_reference_count() const
+	{
+		return this->_reference_count;
+	}
+	void resource::acquire_reference() const
+	{
+		this->_reference_count.fetch_add(1); 
+	}
+	void resource::release_reference() const
+	{
+		if (_reference_count.fetch_sub(1) == 1)
+			delete this;
+	}
 }
