@@ -5,9 +5,9 @@ extern "C"
 {
 	struct log_event;
 
-	FROST_API u64 _stdcall logging_get_log_event_tag();
+	FROST_API u64 _stdcall frost_api_logging_get_log_event_tag();
 
-	FROST_API void _stdcall logging_verbose(
+	FROST_API void _stdcall frost_api_logging_verbose(
 		const wchar_t* message_template,
 		u64 template_length,
 		const wchar_t** params,
@@ -15,7 +15,7 @@ extern "C"
 		u64 param_count,
 		u64 activation_layers);
 
-	FROST_API void _stdcall logging_debug(
+	FROST_API void _stdcall frost_api_logging_debug(
 		const wchar_t* message_template,
 		u64 template_length,
 		const wchar_t** params,
@@ -23,7 +23,7 @@ extern "C"
 		u64 param_count,
 		u64 activation_layers);
 
-	FROST_API void _stdcall logging_info(
+	FROST_API void _stdcall frost_api_logging_info(
 		const wchar_t* message_template,
 		u64 template_length,
 		const wchar_t** params,
@@ -31,7 +31,7 @@ extern "C"
 		u64 param_count,
 		u64 activation_layers);
 
-	FROST_API void _stdcall logging_warning(
+	FROST_API void _stdcall frost_api_logging_warning(
 		const wchar_t* message_template,
 		u64 template_length,
 		const wchar_t** params,
@@ -39,7 +39,7 @@ extern "C"
 		u64 param_count,
 		u64 activation_layers);
 
-	FROST_API void _stdcall logging_error(
+	FROST_API void _stdcall frost_api_logging_error(
 		const wchar_t* message_template,
 		u64 template_length,
 		const wchar_t** params,
@@ -47,7 +47,7 @@ extern "C"
 		u64 param_count,
 		u64 activation_layers);
 
-	FROST_API void _stdcall logging_critical(
+	FROST_API void _stdcall frost_api_logging_critical(
 		const wchar_t* message_template,
 		u64 template_length,
 		const wchar_t** params,
@@ -72,4 +72,11 @@ extern "C"
 		u64				thread_id;
 		u8				level;
 	};
+
+	constexpr u8 logging_level_verbose	= 1;
+	constexpr u8 logging_level_debug	= 2;
+	constexpr u8 logging_level_info		= 4;
+	constexpr u8 logging_level_warning	= 8;
+	constexpr u8 logging_level_error	= 16;
+	constexpr u8 logging_level_critical	= 32;
 }
